requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
Google-api-python-client
google-auth
google-auth-oauthlib
jinja2
python-multipart
python-jobspy
pandas
passlib[bcrypt]
python-jose[cryptography]

